Fix miri-revealed pointer problems.
A few methods were technically creating two mutable references to the
same value at the same time.  The compiler didn't flag it as an error
because they went through raw pointers, and they didn't cause any
runtime problems because the second reference was always dropped before
the first got used, but the fix is easy enough.